divided add event into component
