+++ src/App.tsx
@@ -1,6 +1,5 @@
 import React, { useState, useReducer } from "react";
 import styled from "styled-components";
-import Modal from "react-responsive-modal";
 
 import { StateProvider } from "./state";
 
@@ -21,7 +20,6 @@ import { object } from "prop-types";
 
 export const CurrentDateContext = React.createContext(new Date());
 
-const getEvents = () => JSON.parse(localStorage.getItem("events") || "[]");
 const initialState = { date: new Date() };
 
 const reducer = (state: any, action: any) => {
@@ -42,48 +40,18 @@ const reducer = (state: any, action: any) => {
 };
 
 export const App = () => {
-  const [events, setEvents] = useState(getEvents());
   const [date, setDate] = useState(getToday());
   const [view, setView] = useState(Views.DAY);
-  const [open, setOpen] = useState(false);
-  const [eventValue, setEventValue] = useState({
-    title: "",
-    startTime: "",
-    startDate: "",
-    endTime: "",
-    endDate: ""
-  });
   const handleNavigate = (to: Navigate) => {
     if (to === Navigate.TODAY) setDate(getToday());
     if (to === Navigate.NEXT) setDate(getNextDay(date));
     if (to === Navigate.PREVIOUS) setDate(getPreviousDay(date));
-    if (to === Navigate.CREATE) setOpen(true);
   };
   const handleChangeView = (view: Views) => {
     setView(view);
   };
   const handleDateChange = (date: Date) => {
     setDate(date);
-  };
-  const handleEventChange = (e: { target: HTMLInputElement }) => {
-    const name = e.target.name;
-    const value = e.target.value;
-    setEventValue({ ...eventValue, [name]: value });
-  };
-  const handleAddEvent = () => {
-    const event = {
-      title: eventValue.title,
-      startTime: eventValue.startTime,
-      endTime: eventValue.endTime,
-      startDate: eventValue.startDate,
-      endDate: eventValue.endDate
-    };
-    setEvents([...events, event]);
-    localStorage.setItem("events", JSON.stringify([...events, event]));
-    onCloseModal();
-  };
-  const onCloseModal = () => {
-    setOpen(false);
   };
   let calViewHeader, calView;
   if (view === Views.DAY) {
@@ -114,39 +82,6 @@ export const App = () => {
           {calViewHeader}
           <FullCalViewGridWrapper>{calView}</FullCalViewGridWrapper>
         </FullCalView>
-        <Modal open={open} onClose={onCloseModal} center>
-          <EventTitle
-            type="text"
-            placeholder="Добавьте название"
-            name="title"
-            onChange={e => handleEventChange(e)}
-          />
-          <EventDate
-            name="startDate"
-            type="date"
-            onChange={e => handleEventChange(e)}
-          ></EventDate>
-          <EventTime
-            name="startTime"
-            type="time"
-            onChange={e => handleEventChange(e)}
-          ></EventTime>
-          <EventTime
-            name="endTime"
-            type="time"
-            onChange={e => handleEventChange(e)}
-          ></EventTime>
-          <EventDate
-            name="endDate"
-            type="date"
-            onChange={e => handleEventChange(e)}
-          ></EventDate>
-          <EventSave
-            type="button"
-            value="Сохранить"
-            onClick={() => handleAddEvent()}
-          ></EventSave>
-        </Modal>
       </StyledCalendar>
     </StateProvider>
   );
@@ -157,56 +92,4 @@ const StyledCalendar = styled.div`
   display: grid;
   grid-template-rows: 60px auto;
   grid-template-columns: 260px auto;
-`;
-
-const EventTitle = styled.input`
-  width: 80%;
-  height: 28px;
-  font-size: 22px;
-  color: #3c4043;
-  border: none;
-  border-bottom: 1px solid #dadce0;
-  margin-top: 30px;
-  padding: 5px;
-  outline: none;
-
-  &:focus {
-    border-bottom: 1px solid #4285f4;
-  }
-`;
-
-const EventDate = styled.input`
-  border: none;
-  border-bottom: 1px solid #dadce0;
-  margin-top: 30px;
-  outline: none;
-  margin-left: 10px;
-
-  &:focus {
-    border-bottom: 1px solid #4285f4;
-  }
-`;
-
-const EventTime = styled.input`
-  border: none;
-  border-bottom: 1px solid #dadce0;
-  margin-top: 30px;
-  margin-left: 10px;
-  outline: none;
-
-  &:focus {
-    border-bottom: 1px solid #4285f4;
-  }
-`;
-
-const EventSave = styled.input`
-  width: 122px;
-  height: 37px;
-  color: white;
-  font-size: 14px;
-  background-color: #1a73e8;
-  padding: 0 24px;
-  border: none;
-  border-radius: 4px;
-  cursor: pointer;
 `;
